requirements.txt
requests
oauthlib
httpx
python-dotenv